add alternative join;
