+++ lib/rome/core/rows/rows.py
@@ -4,7 +4,6 @@ import logging
 import time
 import uuid
 
-import pandas as pd
 from sqlalchemy.sql.expression import BinaryExpression
 from sqlalchemy.util._collections import KeyedTuple
 
@@ -12,10 +11,9 @@ from lib.rome.core.dataformat import get_decoder
 from lib.rome.core.lazy import LazyValue
 from lib.rome.core.utils import get_objects, is_novabase
 
-from lib.rome.core.models import get_model_classname_from_tablename
+from tuples import default_panda_building_tuples as simple_building_tuples
+from tuples import sql_panda_building_tuples as building_tuples
 
-import math
-import re
 
 file_logger_enabled = False
 try:
@@ -134,215 +132,6 @@ def extract_joining_criterion_from_relationship(rel, local_table):
     remote_tabledata = {"table": rel.remote_object_tablename, "column": rel.remote_object_field}
     return [local_tabledata, remote_tabledata]
 
-def correct_boolean_int(expression_str):
-    expression_str = expression_str.replace("___deleted == 0", "___deleted != 1")
-    return expression_str
-
-def correct_expression_containing_none(expression_str):
-    # Use trick found here: http://stackoverflow.com/questions/26535563/querying-for-nan-and-other-names-in-pandas
-    terms = expression_str.split()
-    clean_expression = " ".join(terms)
-    if len(terms) > 0:
-        variable_name = terms[0]
-        variable_name = variable_name.replace("(", "")
-        if "is not None" in clean_expression:
-            clean_expression = "(%s == %s)" % (variable_name, variable_name, variable_name)
-        if "is None" in clean_expression:
-            clean_expression = "(%s != %s)" % (variable_name, variable_name, variable_name)
-    return clean_expression
-
-def drop_y(df):
-    # list comprehension of the cols that end with '_y'
-    to_drop = [x for x in df if x.endswith('_y')]
-    df.drop(to_drop, axis=1, inplace=True)
-
-def rename_x(df):
-    cols = list(df.columns)
-    fixed_columns = map(lambda x: re.sub("_x$", "", x), cols)
-    df.columns = fixed_columns
-
-def building_tuples(lists_results, labels, criterions, hints=[]):
-
-    """ Build tuples (join operator in relational algebra). """
-
-    """ Create the Dataframe indexes. """
-    dataframes = []
-    dataindex = {}
-    substitution_index = {}
-    normal_keys_index = {}
-    refactored_keys_index = {}
-    normal_keys_to_key_index = {}
-    refactored_keys_to_key_index = {}
-    refactored_keys_to_table_index = {}
-    index = 0
-
-
-    classname_index = {}
-    for each in labels:
-        classname_index[each] = get_model_classname_from_tablename(each)
-    # if len(lists_results) == 1:
-    #     return map(lambda x: [x], lists_results[0])
-
-
-    for list_results in lists_results:
-        label = labels[index]
-        if len(list_results) == 0:
-            continue
-        keys = map(lambda x: x, list_results[0]) + ["created_at", "updated_at"]
-
-        dataframe = pd.DataFrame(data=list_results, columns=keys)
-
-        for value in keys:
-            normal_key = "%s.%s" % (label, value)
-            refactored_keys = "%s___%s" % (label, value)
-            refactored_keys_to_table_index[refactored_keys] = label
-            normal_keys_to_key_index[normal_key] = value
-            refactored_keys_to_key_index[refactored_keys] = value
-        normal_keys = map(lambda x: "%s.%s" % (label, x), keys)
-        normal_keys_index[label] = normal_keys
-        refactored_keys = map(lambda x: "%s___%s" % (label, x), keys)
-        refactored_keys_index[label] = refactored_keys
-        for (a, b) in zip(normal_keys, refactored_keys):
-            substitution_index[a] = b
-        dataframe.columns = refactored_keys
-        dataframes += [dataframe]
-
-        """ Index the dataframe and create a reverse index. """
-        dataindex[label] = index
-        index += 1
-
-    """ Collecting joining expressions. """
-    joining_pairs = []
-    non_joining_criterions = []
-    _joining_pairs_str_index = {}
-    _nonjoining_criterions_str_index = {}
-    for criterion in criterions:
-        _joining_pairs = criterion.extract_joining_pairs()
-        _nonjoining_criterions = criterion.extract_nonjoining_criterions()
-
-        _nonjoining_criterions_str = str(_nonjoining_criterions)
-
-        if len(_joining_pairs) > 0:
-            _joining_pairs_str = str(sorted(_joining_pairs[0]))
-            if not _joining_pairs_str in _joining_pairs_str_index:
-                _joining_pairs_str_index[_joining_pairs_str] = 1
-                joining_pairs += _joining_pairs
-        if not _nonjoining_criterions_str in _nonjoining_criterions_str_index:
-            _nonjoining_criterions_str_index[_nonjoining_criterions_str] = 1
-            non_joining_criterions += _nonjoining_criterions
-
-
-
-    """ Construct the resulting rows. """
-    if len(labels) > 1 and len(filter(lambda x: len(x) == 0, lists_results)) > 0:
-        return []
-
-    result = None
-
-    if len(lists_results) > 1:
-        processed_tables = []
-        for joining_pair in joining_pairs:
-            """ Preparing the tables that will be joined. """
-
-
-            attribute_1 = joining_pair[0].strip()
-            attribute_2 = joining_pair[1].strip()
-            tablename_1 = attribute_1.split(".")[0]
-            tablename_2 = attribute_2.split(".")[0]
-
-            if tablename_1 not in dataindex or tablename_2 not in dataindex:
-                return []
-            index_1 = dataindex[tablename_1]
-            index_2 = dataindex[tablename_2]
-            dataframe_1 = dataframes[index_1] if not tablename_1 in processed_tables else result
-            dataframe_2 = dataframes[index_2] if not tablename_2 in processed_tables else result
-
-            refactored_attribute_1 = attribute_1.split(".")[0]+"___"+attribute_1.split(".")[1]
-            refactored_attribute_2 = attribute_2.split(".")[0]+"___"+attribute_2.split(".")[1]
-
-            """ Join the tables. """
-            try:
-                result = pd.merge(dataframe_1, dataframe_2, left_on=refactored_attribute_1, right_on=refactored_attribute_2, how="outer")
-                drop_y(result)
-                rename_x(result)
-            except KeyError:
-                return []
-            """ Update the history of processed tables. """
-            processed_tables += [tablename_1, tablename_2]
-            processed_tables = list(set(processed_tables))
-
-    """ Filtering rows. """
-    if result is None:
-        if len(dataframes) == 0:
-            return []
-        result = dataframes[0]
-
-    for non_joining_criterion in non_joining_criterions:
-        expression_str = non_joining_criterion.raw_expression
-        for value in substitution_index:
-            if value in expression_str:
-                corresponding_key = substitution_index[value]
-                expression_str = expression_str.replace(value, corresponding_key)
-        try:
-            corrected_expression = correct_boolean_int(expression_str)
-            corrected_expression = correct_expression_containing_none(corrected_expression)
-            result = result.query(corrected_expression)
-        except:
-            pass
-
-    """ Building the rows. """
-    rows = []
-    # for index, row in result.iterrows():
-    #     value = row
-    #     row = []
-    #     for label in labels:
-    #         refactored_keys = refactored_keys_index[label] if label in refactored_keys_index else {}
-    #         sub_row = {}
-    #         for refactored_key in refactored_keys:
-    #             raw_key = refactored_keys_to_key_index[refactored_key]
-    #             sub_row[raw_key] = value[refactored_key] if refactored_key in value else None
-    #             if type(sub_row[raw_key]) is float and math.isnan(sub_row[raw_key]):
-    #                 sub_row[raw_key] = None
-    #             sub_row[raw_key] = value[refactored_key] if refactored_key in value else 0
-    #         row += [sub_row]
-    #     rows += [row]
-    # return rows
-    #
-    # rows = []
-    # for value in result.values():
-
-    # def row_to_dict()
-
-    # for index, row in result.iterrows():
-    # for row in result.itertuples():
-    #     value = row
-    columns_indexes = {}
-    label_indexes = {}
-    i = 0
-    for refactored_key in result.columns.values:
-        columns_indexes[refactored_key] = i
-        i += 1
-    i = 0
-    for label in labels:
-        label_indexes[label] = i
-        i += 1
-
-    transposed_result = result.transpose()
-    dict_values = transposed_result.to_dict()
-    for value in dict_values.values():
-        row = map(lambda x: {}, labels)
-        for ci in value:
-            table = refactored_keys_to_table_index[ci]
-            table_index = label_indexes[table]
-            key = refactored_keys_to_key_index[ci]
-            v = value[ci]
-            if type(v) is float and math.isnan(v):
-                v = 0
-            if key == "_metadata_novabase_classname":
-                v = classname_index[table]
-            row[table_index][key] = v
-        rows += [row]
-    return rows
 
 def wrap_with_lazy_value(value, only_if_necessary=True, request_uuid=None):
     if value is None:
@@ -416,7 +205,10 @@ def construct_rows(models, criterions, hints, session=None, request_uuid=None):
     part3_starttime = current_milli_time()
 
     """ Building tuples """
-    tuples = building_tuples(list_results, labels, criterions, hints)
+    if len(list_results) > 1:
+        tuples = building_tuples(list_results, labels, criterions, hints)
+    else:
+        tuples = simple_building_tuples(list_results, labels, criterions, hints)
     part4_starttime = current_milli_time()
 
     """ Filtering tuples (cartesian product) """
